Inital commit
Does not do too much at the moment, but does have a functioning locking system for vehicles. About to write an update for it, but wanted to commit what I have so far, in case I break it lol. Have a look through the code for the commands and stuff, I'll provide more info in the next update.